create SongCollectionTest
